+++ Source/Core/Persistence/PersistenceContracts.cs
@@ -1,5 +1,6 @@
 using Mythos.Framework.Characters;
 using Mythos.Framework.Entities;
+using Mythos.Framework.Information;
 using Mythos.Framework.Npcs;
 using Mythos.Framework.Relationships;
 using Mythos.Framework.Regions;
@@ -61,7 +62,7 @@ public sealed record EntityDomainSnapshot(int Version, IReadOnlyList<EntitySnaps
 public sealed class PersistentWorldState
 {
     public PersistentWorldState(EntityRegistry entities, WorldClock clock, RegionFramework regions,
-        CharacterRegistry characters, NpcFramework npcs, RelationshipFramework relationships)
+        CharacterRegistry characters, NpcFramework npcs, RelationshipFramework relationships, InformationFramework information)
     {
         Entities = entities ?? throw new ArgumentNullException(nameof(entities));
         Clock = clock ?? throw new ArgumentNullException(nameof(clock));
@@ -69,6 +70,7 @@ public sealed class PersistentWorldState
         Characters = characters ?? throw new ArgumentNullException(nameof(characters));
         Npcs = npcs ?? throw new ArgumentNullException(nameof(npcs));
         Relationships = relationships ?? throw new ArgumentNullException(nameof(relationships));
+        Information = information ?? throw new ArgumentNullException(nameof(information));
     }
 
     public EntityRegistry Entities { get; }
@@ -77,6 +79,7 @@ public sealed class PersistentWorldState
     public CharacterRegistry Characters { get; }
     public NpcFramework Npcs { get; }
     public RelationshipFramework Relationships { get; }
+    public InformationFramework Information { get; }
 }
 
 public sealed record PersistenceLoadContext(